requirements.txt
setuptools
pillow
opencv-python